Added the fps filter to obtain a constant frame rate video before filtering
As varibale frame rate broke the filter tree I build (as it is based on the assumtion that the input video has a constant frame rate), set the video to use a constant framerate via the fps filter. This doesn't seem to impeed the performance so I will call it for every video.